+++ packages/liveblocks-react-ui/src/primitives/Composer/index.tsx
@@ -27,11 +27,13 @@ import type {
   PointerEvent,
   SyntheticEvent,
 } from "react";
-import React, {
+import {
   forwardRef,
   useCallback,
   useEffect,
+  useId,
   useImperativeHandle,
+  useLayoutEffect,
   useMemo,
   useRef,
   useState,
@@ -97,10 +99,8 @@ import { isKey } from "../../utils/is-key";
 import { Persist, useAnimationPersist, usePersist } from "../../utils/Persist";
 import { Portal } from "../../utils/Portal";
 import { requestSubmit } from "../../utils/request-submit";
-import { useId } from "../../utils/use-id";
 import { useIndex } from "../../utils/use-index";
 import { useInitial } from "../../utils/use-initial";
-import { useLayoutEffect } from "../../utils/use-layout-effect";
 import { useObservable } from "../../utils/use-observable";
 import { useRefs } from "../../utils/use-refs";
 import { toAbsoluteUrl } from "../Comment/utils";