+++ packages/liveblocks-react/src/__tests__/useHistoryVersions.test.tsx
@@ -10,7 +10,7 @@ import {
   type RestRequest,
 } from "msw";
 import { setupServer } from "msw/node";
-import React, { Suspense } from "react";
+import { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
 import MockWebSocket from "./_MockWebSocket";