+++ packages/liveblocks-react/src/__tests__/useMarkThreadAsUnresolved.test.tsx
@@ -1,7 +1,6 @@
 import { nanoid, Permission } from "@liveblocks/core";
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { setupServer } from "msw/node";
-import React from "react";
 
 import { dummyThreadData } from "./_dummies";
 import MockWebSocket from "./_MockWebSocket";