+++ packages/liveblocks-react/src/__tests__/useUser.test.tsx
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import type { ResolveUsersArgs } from "@liveblocks/core";
 import { nanoid } from "@liveblocks/core";
 import { renderHook, screen, waitFor } from "@testing-library/react";
-import React, { Suspense } from "react";
+import { Suspense } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
 import { act, createContextsForTest } from "./_utils";